requirements.txt
numpy==1.19.5
pandas==1.3.3
seaborn==0.11.1
matplotlib==3.4.3
gunicorn==20.1.0
Flask==1.1.2
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
Jinja2==2.11.3
Markdown==3.2.2
MarkupSafe==2.0.1
Werkzeug==2.0.3
itsdangerous==2.0.1
Werkzeug==2.0.3
tensorflow==2.3.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
h5py==2.10.0
glob2==0.7
scikit-learn==0.24.1
itsdangerous==1.1.0
jsonify==0.5
pickleshare==0.7.5
Pillow==8.4.0
setuptools==41.4.0
db-sqlite3==0.0.1
tqdm==4.64.0
keras-efficientnets-0.1.7
efficientnet-1.1.1
torch==1.8.1
torchvision==0.9.1
opencv-python==4.1.2.30
opencv-contrib-python==4.1.2.30

####If Jupyter NOtebook is not working then install this packages

#pip install --upgrade jupyterhub
#pip install --upgrade --user nbconvert